Get Activities by time spent grouped by description

+++ src/main/java/com/example/ActualApp/controller/dto/ActivityDto.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 public record ActivityDto (
          UUID id,
          String description,
-         int timeSpentInMinutes,
+         long timeSpentInMinutes,
          LocalDate date,
          boolean completed
 ) {
