+++ src/main/java/com/example/ActualApp/repository/entity/Activity.java
@@ -21,7 +21,7 @@ import java.util.UUID;
 public class Activity {
     private static final int MINUTES_IN_A_DAY = 1440;
 
-    public Activity(String description, int timeSpentInMinutes, @NotNull LocalDate date, boolean completed, ActivityCategory category) {
+    public Activity(String description, long timeSpentInMinutes, @NotNull LocalDate date, boolean completed, ActivityCategory category) {
         this.description = description;
         this.timeSpentInMinutes = timeSpentInMinutes;
         this.date = date;
@@ -35,7 +35,7 @@ public class Activity {
     @NotBlank
     private String description;
     @Range(min = 0, max = MINUTES_IN_A_DAY, message = "Time spent should not be less than 0 and greater than 1440")
-    private int timeSpentInMinutes;
+    private long timeSpentInMinutes;
     @JsonBackReference
     @ManyToOne
     private ActivityCategory category;
